+++ oscar-cbls/src/main/scala/oscar/cbls/algo/magicArray/MagicIntArrayStacked.scala
@@ -16,20 +16,20 @@ package oscar.cbls.algo.magicArray
   ******************************************************************************/
 
 
-class MagicIntArrayStacked(maxLevel:Int, initVal:(Int => Long), size:Int) extends Iterable[Long]{
+class MagicIntArrayStacked(maxLevel:Int, initVal:(Int => Int), size:Int) extends Iterable[Int]{
 
-  private[this] val levelToArray:Array[Array[Long]] = Array.tabulate(maxLevel+1)(level => if(level == 0) Array.tabulate(size)(initVal) else Array.fill(size)(0L))
+  private[this] val levelToArray:Array[Array[Int]] = Array.tabulate(maxLevel+1)(level => if(level == 0) Array.tabulate(size)(initVal) else Array.fill(size)(0))
   private[this] val levelToIsValueChangedAtNextLevel:Array[IterableMagicBoolArray] = Array.tabulate(maxLevel)(level => new IterableMagicBoolArray(size,false))
   private[this] var currentLevel:Int = 0
 
   def level:Int = currentLevel - 1
 
-  def update(indice:Int,value:Long){
+  def update(indice:Int,value:Int){
     levelToArray(currentLevel)(indice) = value
     if(currentLevel!=0)levelToIsValueChangedAtNextLevel(currentLevel-1)(indice) = true
   }
 
-  def apply(indice:Int):Long = {
+  def apply(indice:Int):Int = {
     var attemptLevel = currentLevel
     while(attemptLevel>0){
       val levelBelow = attemptLevel - 1
@@ -63,11 +63,11 @@ class MagicIntArrayStacked(maxLevel:Int, initVal:(Int => Long), size:Int) extend
     }
   }
 
-  def cloneTopArray:Array[Long] = {
+  def cloneTopArray:Array[Int] = {
     Array.tabulate(size)(this(_))
   }
 
-  override def iterator : Iterator[Long] = {
+  override def iterator : Iterator[Int] = {
     cloneTopArray.iterator
   }
