+++ oscar-cbls/src/main/scala/oscar/cbls/business/routing/neighborhood/RouteExchange.scala
@@ -3,7 +3,6 @@ package oscar.cbls.business.routing.neighborhood
 import oscar.cbls.algo.search.HotRestart
 import oscar.cbls.business.routing.model.VRP
 import oscar.cbls.core.search.{EasyNeighborhoodMultiLevel, First, LoopBehavior}
-import oscar.cbls._
 
 
 /**
@@ -11,8 +10,8 @@ import oscar.cbls._
  * It does not modifies the routes themselves. It just exchanges the vehicles
  * @author renaud.delandtsheer@cetic.be
  */
-case class RouteExchange(firstVehicles:()=>Iterable[Long],
-                         secondVehicles:()=>Long=>Iterable[Long],
+case class RouteExchange(firstVehicles:()=>Iterable[Int],
+                         secondVehicles:()=>Int=>Iterable[Int],
                          vrp:VRP,
                          neighborhoodName: String = "RouteExchange",
                          selectFirstVehicleBehavior:LoopBehavior = First(),
@@ -27,18 +26,18 @@ case class RouteExchange(firstVehicles:()=>Iterable[Long],
   val n = vrp.n
 
   //the indice to start with for the exploration
-  var startFirstVehicle: Long = 0
+  var startFirstVehicle: Int = 0
 
-  var firstVehicle:Long = -1L
-  var secondVehicle:Long = -1L
-  var positionOfFirstVehicle:Long = -1L
-  var positionOfVehicleNextToFirst:Long = -1L
+  var firstVehicle:Int = -1
+  var secondVehicle:Int = -1
+  var positionOfFirstVehicle:Int = -1
+  var positionOfVehicleNextToFirst:Int = -1
   var firstVehicleIsEmpty:Boolean = false
-  var positionOfSecondVehicle:Long = -1L
-  var positionOfVehicleNextToSecond:Long = -1L
+  var positionOfSecondVehicle:Int = -1
+  var positionOfVehicleNextToSecond:Int = -1
   var secondVehicleIsEmpty:Boolean = false
 
-  override def exploreNeighborhood(initialObj: Long){
+  override def exploreNeighborhood(initialObj: Int){
 
     val iterationSchemeOnZone =
       if (hotRestart) HotRestart(firstVehicles(), startFirstVehicle)
@@ -46,7 +45,7 @@ case class RouteExchange(firstVehicles:()=>Iterable[Long],
 
     val startValue = seq.defineCurrentValueAsCheckpoint(true)
 
-    def evalObjAndRollBack() : Long = {
+    def evalObjAndRollBack() : Int = {
       val a = obj.value
       seq.rollbackToTopCheckpoint(startValue)
       a
@@ -61,8 +60,8 @@ case class RouteExchange(firstVehicles:()=>Iterable[Long],
       require(firstVehicle < v,"first vehicle is not <v:" + firstVehicle)
 
       positionOfFirstVehicle = startValue.positionOfAnyOccurrence(firstVehicle).get
-      positionOfVehicleNextToFirst = if(firstVehicle == v-1L) startValue.size else startValue.positionOfAnyOccurrence(firstVehicle+1L).get
-      firstVehicleIsEmpty = positionOfFirstVehicle+1L == positionOfVehicleNextToFirst
+      positionOfVehicleNextToFirst = if(firstVehicle == v-1) startValue.size else startValue.positionOfAnyOccurrence(firstVehicle+1).get
+      firstVehicleIsEmpty = positionOfFirstVehicle+1 == positionOfVehicleNextToFirst
 
       if(!skipFirstVehicleIfEmptyRoute || !firstVehicleIsEmpty){
 
@@ -75,8 +74,8 @@ case class RouteExchange(firstVehicles:()=>Iterable[Long],
           if(!breakSymmetriesAmongVehicles || secondVehicle > firstVehicle){
 
             positionOfSecondVehicle = startValue.positionOfAnyOccurrence(secondVehicle).get
-            positionOfVehicleNextToSecond = if(secondVehicle == v-1L) startValue.size else startValue.positionOfAnyOccurrence(secondVehicle+1L).get
-            secondVehicleIsEmpty = positionOfSecondVehicle +1L == positionOfVehicleNextToSecond
+            positionOfVehicleNextToSecond = if(secondVehicle == v-1) startValue.size else startValue.positionOfAnyOccurrence(secondVehicle+1).get
+            secondVehicleIsEmpty = positionOfSecondVehicle +1 == positionOfVehicleNextToSecond
 
             if(!firstVehicleIsEmpty || !secondVehicleIsEmpty){
 
@@ -94,51 +93,51 @@ case class RouteExchange(firstVehicles:()=>Iterable[Long],
     }
     seq.releaseTopCheckpoint()
     //For the hotRestart
-    startFirstVehicle = firstVehicle + 1L
+    startFirstVehicle = firstVehicle + 1
 
     //to ensure that we do not access these values without assigning them first, we set a trap value
-    firstVehicle = -1L
-    secondVehicle = -1L
-    positionOfFirstVehicle = -1L
-    positionOfVehicleNextToFirst = -1L
-    positionOfSecondVehicle = -1L
-    positionOfVehicleNextToSecond = -1L
+    firstVehicle = -1
+    secondVehicle = -1
+    positionOfFirstVehicle = -1
+    positionOfVehicleNextToFirst = -1
+    positionOfSecondVehicle = -1
+    positionOfVehicleNextToSecond = -1
   }
 
 
-  override def instantiateCurrentMove(newObj: Long) =
-    RouteExchangeMove(firstVehicle:Long,
-      positionOfFirstVehicle:Long,
-      positionOfVehicleNextToFirst:Long,
+  override def instantiateCurrentMove(newObj: Int) =
+    RouteExchangeMove(firstVehicle:Int,
+      positionOfFirstVehicle:Int,
+      positionOfVehicleNextToFirst:Int,
       firstVehicleIsEmpty:Boolean,
-      secondVehicle:Long,
-      positionOfSecondVehicle:Long,
-      positionOfVehicleNextToSecond:Long,
+      secondVehicle:Int,
+      positionOfSecondVehicle:Int,
+      positionOfVehicleNextToSecond:Int,
       secondVehicleIsEmpty:Boolean,
       newObj, this, neighborhoodName)
 
   override def reset(): Unit = {
-    startFirstVehicle = 0L
+    startFirstVehicle = 0
   }
 
-  def doMove(positionOfFirstVehicle:Long,
-             positionOfVehicleNextToFirst:Long,
+  def doMove(positionOfFirstVehicle:Int,
+             positionOfVehicleNextToFirst:Int,
              firstVehicleIsEmpty:Boolean,
-             positionOfSecondVehicle:Long,
-             positionOfVehicleNextToSecond:Long,
+             positionOfSecondVehicle:Int,
+             positionOfVehicleNextToSecond:Int,
              secondVehicleIsEmpty:Boolean) {
     if (firstVehicleIsEmpty) {
       if (secondVehicleIsEmpty) {
         System.err.println("doing Route exchange with two empty routes")
       } else {
-        seq.move(positionOfSecondVehicle + 1L, positionOfVehicleNextToSecond - 1L, positionOfFirstVehicle, false)
+        seq.move(positionOfSecondVehicle + 1, positionOfVehicleNextToSecond - 1, positionOfFirstVehicle, false)
       }
     } else {
       if (secondVehicleIsEmpty) {
-        seq.move(positionOfFirstVehicle + 1L, positionOfVehicleNextToFirst - 1L, positionOfSecondVehicle, false)
+        seq.move(positionOfFirstVehicle + 1, positionOfVehicleNextToFirst - 1, positionOfSecondVehicle, false)
       } else {
-        seq.swapSegments(positionOfFirstVehicle + 1L, positionOfVehicleNextToFirst - 1L, false,
-          positionOfSecondVehicle + 1L, positionOfVehicleNextToSecond - 1L, false)
+        seq.swapSegments(positionOfFirstVehicle + 1, positionOfVehicleNextToFirst - 1, false,
+          positionOfSecondVehicle + 1, positionOfVehicleNextToSecond - 1, false)
       }
     }
   }
@@ -147,21 +146,21 @@ case class RouteExchange(firstVehicles:()=>Iterable[Long],
 /**
  * @author renaud.delandtsheer@cetic.be
  */
-case class RouteExchangeMove(firstVehicle:Long,
-                             positionOfFirstVehicle:Long,
-                             positionOfVehicleNextToFirst:Long,
+case class RouteExchangeMove(firstVehicle:Int,
+                             positionOfFirstVehicle:Int,
+                             positionOfVehicleNextToFirst:Int,
                              firstVehicleIsEmpty:Boolean,
-                             secondVehicle:Long,
-                             positionOfSecondVehicle:Long,
-                             positionOfVehicleNextToSecond:Long,
+                             secondVehicle:Int,
+                             positionOfSecondVehicle:Int,
+                             positionOfVehicleNextToSecond:Int,
                              secondVehicleIsEmpty:Boolean,
-                             override val objAfter: Long,
+                             override val objAfter: Int,
                              override val neighborhood: RouteExchange,
                              override val neighborhoodName: String = "RouteExchangeMove")
   extends VRPSMove(objAfter, neighborhood, neighborhoodName, neighborhood.vrp){
 
   //TODO: Implement this
-  override def impactedPoints: Iterable[Long] = ???
+  override def impactedPoints: Iterable[Int] = ???
 
   override def commit() {
     neighborhood.doMove(positionOfFirstVehicle,
