+++ oscar-cbls/src/main/scala/oscar/cbls/business/routing/neighborhood/VRPSMove.scala
@@ -19,13 +19,13 @@ import oscar.cbls.business.routing.model.VRP
 import oscar.cbls.core.computation.Variable
 import oscar.cbls.core.search.{Move, Neighborhood}
 
-abstract class VRPSMove(override val objAfter: Long,
+abstract class VRPSMove(override val objAfter: Int,
                        val neighborhood:Neighborhood,
                        override val neighborhoodName:String = null, vrp:VRP)
   extends Move(objAfter, neighborhoodName){
 
   override def touchedVariables: List[Variable] = List(vrp.routes)
 
-  def impactedPoints:Iterable[Long]
+  def impactedPoints:Iterable[Int]
 
 }
